requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
gspread==6.1.2
google-auth==2.33.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
requests==2.32.3
tenacity==9.0.0
langchain==0.2.12
langchain-openai==0.1.22
langchain-google-genai==1.0.7
